refactor: do not store fetchWrapper.fetch() via destructuring
By storing a reference to fetchWrapper.fetch() via destructuring, we are
unable to stub out the function later when we attempt to test its
execution and properties [[1](https://stackoverflow.com/a/52591287/10366522)]. By instead storing a reference to the calling
environment, whether it be nodejs or the browser, we can then invoke fetch() at the
point where the request is being made. This relatively small change will
allow us to expand our tests to ensure that the API URL is constructed
correctly, and all default headers are passed.