fix: always stringify error data
APIError is typically passed error data in form of a JSON object. Part
of the error-handling logic in imgix-api.js attempts to parse the error
data from a string, but this logic is generally never hit (it's
arguable whether it is still useful to keep). Regardless of how the
data is received, it is always converted to an object before
stringified in toString(), therefore we can eliminate extra logic for
handling non-string error data.